+++ vncdotool/decoders/copyrect.py
@@ -6,14 +6,14 @@ from typing import ClassVar, Iterator
 
 from ..const import Encoding
 from ..pixelformat import PixelFormat
-from .base import ClientDecoder, Rect
+from .base import ClientDecoder
 
 
 class CopyRectDecoder(ClientDecoder):
     ENCODING: ClassVar[Encoding] = Encoding.COPY_RECTANGLE
 
     def decodeForClient(
-        self, client: object, rect: Rect, pixel_format: PixelFormat
+        self, client: object, rect: tuple[int, int, int, int], pixel_format: PixelFormat
     ) -> Iterator[int]:
         block = yield 4
         srcx, srcy = unpack("!HH", block)